refactor(examples): standardize actor registry filenames
