+++ examples/flue/agents/assistant.ts
@@ -1,6 +1,6 @@
 import { createAgent } from "@flue/runtime";
 import { agentOSSandbox } from "@rivet-dev/agentos-flue";
-import { registry } from "../registry.js";
+import { registry } from "../actors.js";
 
 export default createAgent(() => ({
 	model: "anthropic/claude-sonnet-4-6",
