+++ examples/vercel-eve/agent/sandbox.ts
@@ -1,6 +1,6 @@
 import { agentOSBackend } from "@rivet-dev/agentos-eve";
 import { defineSandbox } from "eve/sandbox";
-import { registry } from "../registry";
+import { registry } from "../actors";
 
 export default defineSandbox({
 	backend: agentOSBackend({ actor: "vm", registry }),
