+++ examples/vercel-eve/world.ts
@@ -1,4 +1,4 @@
 import { createWorld as createRivetWorld } from "@rivet-dev/vercel-world";
-import { registry } from "./registry";
+import { registry } from "./actors";
 
 export const createWorld = () => createRivetWorld({ registry });
