+++ website/src/content/docs/docs/frameworks/vercel-eve.mdx
@@ -79,7 +79,7 @@ Create `world.ts`:
 
 ```ts title="world.ts"
 import { createWorld as createRivetWorld } from "@rivet-dev/vercel-world";
-import { registry } from "./registry";
+import { registry } from "./actors";
 
 export const createWorld = () => createRivetWorld({ registry });
 ```
@@ -91,9 +91,9 @@ be ready.
 
 <Step title="Configure agentOS">
 
-Create `registry.ts`:
+Create `actors.ts`:
 
-```ts title="registry.ts"
+```ts title="actors.ts"
 import { agentOS, setup } from "@rivet-dev/agentos";
 import { vercelWorldActors } from "@rivet-dev/vercel-world/registry";
 
@@ -114,7 +114,7 @@ Create `agent/sandbox.ts`:
 ```ts title="agent/sandbox.ts"
 import { agentOSBackend } from "@rivet-dev/agentos-eve";
 import { defineSandbox } from "eve/sandbox";
-import { registry } from "../registry";
+import { registry } from "../actors";
 
 export default defineSandbox({
 	backend: agentOSBackend({ actor: "vm", registry }),
